refactor: rename pre-SRoC scenario files to presroc-licence- prefix
Applies the new licence-type-prefix naming convention to the two scenario files this PR touches, and updates the two spec files that import the renamed charge-version scenario.
