+++ tests/support/scenarios/presroc-licence.scenario.js
